+++ src/tooling/piral-cli/src/injectors/pilet-injector.ts
@@ -1,5 +1,5 @@
 import { URL } from 'url';
-import { join } from 'path';
+import { join, resolve } from 'path';
 import { EventEmitter } from 'events';
 import { readFile, stat, writeFile } from 'fs/promises';
 import { KrasInjector, KrasRequest, KrasInjectorConfig, KrasConfiguration, KrasResult } from 'kras';
@@ -140,15 +140,18 @@ export default class PiletInjector implements KrasInjector {
       const cbs = {};
 
       if (app.endsWith('/app')) {
-        const packageJson = require(`${app}/../package.json`);
+        try {
+          const path = resolve(app, '..', 'package.json');
+          const packageJson = require(path);
 
-        if (typeof packageJson.piralCLI.source === 'string') {
-          this.proxyInfo = {
-            source: packageJson.piralCLI.source,
-            files: packageJson.files,
-            date: new Date(packageJson.piralCLI.timestamp),
-          };
-        }
+          if (typeof packageJson.piralCLI.source === 'string') {
+            this.proxyInfo = {
+              source: packageJson.piralCLI.source,
+              files: packageJson.files,
+              date: new Date(packageJson.piralCLI.timestamp),
+            };
+          }
+        } catch {}
       }
 
       core.on('user-connected', (e) => {